added course app; added add course functionality; formatted classlist table; added classlist link to navbar
