+++ course/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from course.models import Course, Session
+
+# Register your models here.
+admin.site.register(Course)
+admin.site.register(Session)
