+++ course/templates/course/add_course.html
@@ -1,0 +1,15 @@
+{% extends 'base.html' %}
+{% block content %}
+<form method="post" action="{% url 'add_course' %}">
+
+    {% csrf_token %}
+
+    {% for field in form  %}
+        {{field.label}}
+        {{field}}
+        </br>
+    {% endfor %}
+    <input type="submit" value="Add"/>
+    <input type="hidden" value="{{next}}" name="next"/>
+</form>
+{% endblock %}
