+++ mbu/admin.py
@@ -1,8 +1,6 @@
 from django.contrib import admin
-from mbu.models import Course, Session, Scout, Troop, TroopContact, Address, Counselor, Venture, MeritBadgeUniversity, WaitingList
+from mbu.models import Scout, Troop, TroopContact, Address, Counselor, Venture, MeritBadgeUniversity, WaitingList
 
-admin.site.register(Course)
-admin.site.register(Session)
 admin.site.register(Scout)
 admin.site.register(Troop)
 admin.site.register(TroopContact)
