+++ mbu/templates/mbu/classlist.html
@@ -1,10 +1,12 @@
-<!doctype html>
-<html>
-<head>
-    <title>Class Listing</title>
-</head>
-<body>
-    <table class="table">
+{% extends "base.html" %}
+
+{% block title %}
+    Class Listing
+{% endblock %}
+
+{% block content %}
+<table class="table">
+	<thead>
         <tr>
             <td>ID</td>
             <td>Name</td>
@@ -12,7 +14,9 @@
             <td>Location</td>
             <td>Teacher</td>
             <td>Requirements</td>
-        </tr>
+   		</tr>
+   	</thead>
+   	<tbody class="table-striped">
         {% for class in classlist %}
         <tr>
             <td>{{ class.id }}</td>
@@ -23,6 +27,7 @@
             <td><a href="{% url class.link class.id %}">Link</a></td>
         </tr>
         {% endfor %}
-    </table>
-</body>
-</html>
+    </tbody>
+</table>
+
+{% endblock %}
