+++ utmbu/settings.py
@@ -39,7 +39,8 @@ INSTALLED_APPS = (
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'mbu'
+    'mbu',
+    'course'
 )
 
 MIDDLEWARE_CLASSES = (
