+++ utmbu/urls.py
@@ -26,3 +26,9 @@ urlpatterns += patterns(
     url(r'^scout/register/', 'mbu.views.register_scout', name = 'register_scout'),
     url(r'^scoutmaster/', 'mbu.views.view_scoutmaster_page', name = 'scoutmaster'),
 )
+
+#Course URLs
+urlpatterns += patterns(
+    '',
+    url(r'^course/add', 'course.views.add_course', name = 'add_course')
+)
